Extract JS and CSS for subnav component from circleci.com
This CSS file depends on the definition of several site-wide variables before
inclusion, but we've chosen not to include those here so that users of this
component can declare their own values for them.

We've also chosen not to extract the HTML template for this component given
that the HTML is less complex and doesn't benefit as much from deduplication
across projects.