Allow user of subnav.js to trigger a state refresh
This allows a page including subnav.js to trigger an event to refresh the
active anchor and positioning of the subnav.

This opens up use cases where the subnav is not shown upon page load, but shown
later after load. Without this, in that case the position and active anchor
calculations yield incorrect results because the positions of all hidden
elements are 0.