Support multiple subnavs on the same page
This change enables the subnav component to work correctly when multiple subnav
components are present on the same page. This allows AB testing of multiple
versions of a subnav.

NOTE: This does include a potentially-breaking change for dependent products.
The means of identifying the subnav placeholder element has been switched from
an ID to a class (to allow multiple copies on the same page).

If your project markup, styles, or otherwise depends on `#subnav-placeholder`,
you will instead need to refer to `.subnav-placeholder`.